+++ R/theme-inter.R
@@ -65,7 +65,7 @@ theme_inter <- function(
   strip_text_size = 12,
   strip_placement = "outer",
   caption_size = 9,
-  caption_margin = 8,
+  caption_margin = 11,
   axis_text_size = base_size,
   axis_title_size = 9,
   axis_title_just = "rt",