+++ NEWS.md
@@ -1,2 +1,6 @@
+0.2.2
+* title/subtitle/caption are now `ggtext::element_textbox_simple()`
+* minor margin tweaks
+
 0.1.0 
 * Initial release